+++ src/lib/functions/mod.rs
@@ -7,6 +7,7 @@
 #[path = "./mod_test.rs"]
 mod mod_test;
 
+mod decode_func;
 mod remove_empty_func;
 mod split_func;
 mod trim_func;
@@ -14,10 +15,16 @@ mod trim_func;
 use crate::types::{Step, Task};
 
 fn run_function(function_name: &str, function_args: &Vec<String>) -> Vec<String> {
+    debug!(
+        "Running function: {} arguments: {:#?}",
+        &function_name, &function_args
+    );
+
     match function_name {
         "split" => split_func::invoke(function_args),
         "remove-empty" => remove_empty_func::invoke(function_args),
         "trim" => trim_func::invoke(function_args),
+        "decode" => decode_func::invoke(function_args),
         _ => {
             error!("Unknown function: {}", &function_name);
             panic!("Unknown function: {}", &function_name);