+++ src/lib/functions/remove_empty_func_test.rs
@@ -4,18 +4,18 @@ use envmnt;
 
 #[test]
 #[should_panic]
-fn invoke_empty() {
+fn remote_empty_invoke_empty() {
     invoke(&vec![]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_too_many_args() {
+fn remote_empty_invoke_invalid_too_many_args() {
     invoke(&vec!["TEST".to_string(), "1".to_string()]);
 }
 
 #[test]
-fn invoke_exists_with_value() {
+fn remote_empty_invoke_exists_with_value() {
     envmnt::set("TEST_REMOVE_EMPTY_VALID", "abc");
 
     let output = invoke(&vec!["TEST_REMOVE_EMPTY_VALID".to_string()]);
@@ -24,7 +24,7 @@ fn invoke_exists_with_value() {
 }
 
 #[test]
-fn invoke_exists_empty() {
+fn remote_empty_invoke_exists_empty() {
     envmnt::set("TEST_REMOVE_EMPTY_EMPTY", "");
 
     let output = invoke(&vec!["TEST_REMOVE_EMPTY_EMPTY".to_string()]);
@@ -33,7 +33,7 @@ fn invoke_exists_empty() {
 }
 
 #[test]
-fn invoke_not_exists() {
+fn remote_empty_invoke_not_exists() {
     let output = invoke(&vec!["TEST_REMOVE_EMPTY_NOT_EXISTS".to_string()]);
 
     assert_eq!(output.len(), 0);