+++ src/lib/functions/split_func_test.rs
@@ -4,30 +4,30 @@ use envmnt;
 
 #[test]
 #[should_panic]
-fn invoke_empty() {
+fn split_invoke_empty() {
     invoke(&vec![]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_too_many_args() {
+fn split_invoke_invalid_too_many_args() {
     invoke(&vec!["TEST".to_string(), "1".to_string(), "2".to_string()]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_split_by_big() {
+fn split_invoke_invalid_split_by_big() {
     invoke(&vec!["TEST".to_string(), "ab".to_string()]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_split_by_empty() {
+fn split_invoke_invalid_split_by_empty() {
     invoke(&vec!["TEST".to_string(), "".to_string()]);
 }
 
 #[test]
-fn invoke_exists_splitted_comma() {
+fn split_invoke_exists_splitted_comma() {
     envmnt::set("TEST_SPLIT_VALUE", "1,2,3,4");
 
     let output = invoke(&vec!["TEST_SPLIT_VALUE".to_string(), ",".to_string()]);
@@ -36,7 +36,7 @@ fn invoke_exists_splitted_comma() {
 }
 
 #[test]
-fn invoke_exists_splitted_space() {
+fn split_invoke_exists_splitted_space() {
     envmnt::set("TEST_SPLIT_VALUE", "1 2 3 4");
 
     let output = invoke(&vec!["TEST_SPLIT_VALUE".to_string(), " ".to_string()]);
@@ -45,7 +45,7 @@ fn invoke_exists_splitted_space() {
 }
 
 #[test]
-fn invoke_exists_not_splitted() {
+fn split_invoke_exists_not_splitted() {
     envmnt::set("TEST_SPLIT_VALUE", "1,2,3,4");
 
     let output = invoke(&vec!["TEST_SPLIT_VALUE".to_string(), "|".to_string()]);
@@ -54,7 +54,7 @@ fn invoke_exists_not_splitted() {
 }
 
 #[test]
-fn invoke_not_exists() {
+fn split_invoke_not_exists() {
     let output = invoke(&vec![
         "TEST_SPLIT_VALUE_NOT_EXISTS".to_string(),
         ",".to_string(),