+++ src/lib/functions/trim_func_test.rs
@@ -4,24 +4,24 @@ use envmnt;
 
 #[test]
 #[should_panic]
-fn invoke_empty() {
+fn trim_invoke_empty() {
     invoke(&vec![]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_too_many_args() {
+fn trim_invoke_invalid_too_many_args() {
     invoke(&vec!["TEST".to_string(), "1".to_string(), "2".to_string()]);
 }
 
 #[test]
 #[should_panic]
-fn invoke_invalid_trim_type() {
+fn trim_invoke_invalid_trim_type() {
     invoke(&vec!["TEST".to_string(), "bad".to_string()]);
 }
 
 #[test]
-fn invoke_exists_with_value() {
+fn trim_invoke_exists_with_value() {
     envmnt::set("TEST_TRIM_VALID", "abc");
 
     let output = invoke(&vec!["TEST_TRIM_VALID".to_string()]);
@@ -30,7 +30,7 @@ fn invoke_exists_with_value() {
 }
 
 #[test]
-fn invoke_exists_empty() {
+fn trim_invoke_exists_empty() {
     envmnt::set("TEST_TRIM_EMPTY", "");
 
     let output = invoke(&vec!["TEST_TRIM_EMPTY".to_string()]);
@@ -39,14 +39,14 @@ fn invoke_exists_empty() {
 }
 
 #[test]
-fn invoke_not_exists() {
+fn trim_invoke_not_exists() {
     let output = invoke(&vec!["TEST_TRIM_NOT_EXISTS".to_string()]);
 
     assert_eq!(output.len(), 0);
 }
 
 #[test]
-fn invoke_all_spaces() {
+fn trim_invoke_all_spaces() {
     envmnt::set("TEST_TRIM_ALL_SPACES", "");
 
     let output = invoke(&vec!["TEST_TRIM_ALL_SPACES".to_string()]);
@@ -55,7 +55,7 @@ fn invoke_all_spaces() {
 }
 
 #[test]
-fn invoke_partial_spaces() {
+fn trim_invoke_partial_spaces() {
     envmnt::set("TEST_TRIM_ALL_PARTIAL_SPACES", "   123   123   ");
 
     let output = invoke(&vec!["TEST_TRIM_ALL_PARTIAL_SPACES".to_string()]);
@@ -64,7 +64,7 @@ fn invoke_partial_spaces() {
 }
 
 #[test]
-fn invoke_trim_start() {
+fn trim_invoke_trim_start() {
     envmnt::set("TEST_TRIM_ALL_PARTIAL_SPACES", "   123   ");
 
     let output = invoke(&vec![
@@ -76,7 +76,7 @@ fn invoke_trim_start() {
 }
 
 #[test]
-fn invoke_trim_end() {
+fn trim_invoke_trim_end() {
     envmnt::set("TEST_TRIM_ALL_PARTIAL_SPACES", "   123   ");
 
     let output = invoke(&vec![